pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.1.0"
description = "MLOps project for pairwise binary classification"
authors = ["Alvar Khairullin <alvkha@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.0.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
hydra-core = "^1.3.0"
omegaconf = "^2.3.0"
dvc = "^3.0.0"
pyyaml = "^6.0"
matplotlib = "^3.7.0"
gitpython = "^3.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "venv",
    ".venv",
    "build",
    "dist",
]

